Finalize
* Move notFound() test override from RestProblemSpringBoot3IT
  to AbstractRestProblemSpringBootIT
* Merge belgif-rest-problem-spring documentation into
  belgif-rest-problem-spring-boot-3 documentation
* Remove activeByDefault=false (because that's the default)
